% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.auc.R
\name{bart.auc}
\alias{bart.auc}
\title{AUC for BART object}
\usage{
bart.auc(model)
}
\arguments{
\item{model}{A BART model object generated by the dbarts package}
}
\description{
An all-purpose model diagnostic, that mostly generates an AUC value and receiver operator curve. It also prints the optimal threshold based on the true skill statistic (and return the associated Type I and II error rates).
}
