% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.map.R
\name{bart.map}
\alias{bart.map}
\title{Spatial prediction}
\usage{
bart.map(model, inputstack, ci = FALSE, plots = FALSE)
}
\arguments{
\item{model}{A model object from dbarts package}

\item{inputstack}{A RasterStack of environmental predictors; make sure they have the same names as the predictors that went into the model object. (If they don't it'll throw an error.)}

\item{ci}{Return 5\% and 95\% credible interval bounds on the predictions, also as rasters. (Returns a list of mean, upper, and lower if you turn this on.) Note that this will make it take much longer to run.}

\item{plots}{Automatically produce a plot of the mean, the 95\% CI width, the lower and the upper bound.}
}
\description{
Spatial prediction of a BART model
}
