% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.step.R
\name{bart.step}
\alias{bart.step}
\title{Full-service variable selection}
\usage{
bart.step(x.data, y.data, iter.step = 100, tree.step = 10,
  iter.plot = 100, full = FALSE, quiet = FALSE)
}
\arguments{
\item{x.data}{A data frame of covariates}

\item{y.data}{A vector of outcomes (1/0)}

\item{iter.step}{How many BART models to run for each iteration of the stepwise reduction}

\item{tree.step}{How many trees to use in the variable set reduction.Should be a SMALL number (10 or 20 trees) in order to create the maximum disparity in variable importance between informative and uninformative predictors (recommendations taken from Chipman et al. 2010).}

\item{iter.plot}{How many iterations to use in the first diagnostic plot}

\item{full}{If this is set to FALSE (by default), this runs a stepwise variable set reduction and returns a model with the optimal variable step - much like gbm::gbm.step() or similar functions. In running varimp.step() it generates a single plot of RMSE against variables dropped. If this is set to TRUE, it also runs summary() on the model, and two additional plots are generated: the initial variable importance diagnostic generated by varimp.diag() (this is SLOW), and a final variable importance bar chart for the final model.}
}
\value{
Returns a model object run with the optimal, reduced variable set.
}
\description{
A wrapper for a few core functions, including a few diagnostic plots of variable importance, and the automated stepwise variable set reduction algorithm.
}
