% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.var.R
\name{bart.var}
\alias{bart.var}
\title{Full-service variable selection}
\usage{
bart.var(xdata, ydata, iter.step = 100, tree.step = 10,
  iter.plot = 100)
}
\arguments{
\item{xdata}{A data frame of covariates}

\item{ydata}{A vector of outcomes (1/0)}

\item{iter.step}{How many BART models to run for each iteration of the stepwise reduction}

\item{tree.step}{How many trees to use in the variable set reduction}

\item{iter.plot}{How many iterations to use in the first diagnostic plot}
}
\value{
Returns a list of (1) Variables (the optimal variable set) and (2) Model.object, the best model using those
}
\description{
A wrapper for a few core functions, including two diagnostic plots of variable importance and the automated stepwise variable set reduction algorithm. Returns an object with the model and the list of the best variables
}
