% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigstack.R
\name{bigstack}
\alias{bigstack}
\title{Big stack}
\usage{
bigstack(stack, by, quiet = FALSE)
}
\arguments{
\item{stack}{A raster stack. It can be a one-layer stack if you'd like, but then just use velox's $aggregate function.}

\item{by}{The factor by which you want to aggregate your rasterstack.}
}
\description{
Big stack is an incredibly goofy convenience tool (analogous to raster::aggregate) I use that runs a velox wrapper to aggregate up a rasterstack. This is useful if you're doing high-res mapping and you want to do some predictive checks.
}
