% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embarcadero-package.R
\docType{package}
\name{embarcadero-package}
\alias{embarcadero-package}
\alias{embarcadero}
\title{Species distribution modeling with BART}
\description{
This package is a wrapper for the `dbarts` package for Bayesian additive regression trees, which is designed to make it easier to do species distribution models with BART.
}
\author{
Colin Carlson (\email{cjc322@georgetown.edu})
}
