% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{partial}
\alias{partial}
\title{Better, nicer, friendlier partials}
\usage{
partial(model, x.vars = NULL, equal = FALSE, smooth = 1,
  posterior = NULL)
}
\arguments{
\item{model}{A dbarts model object}

\item{equal}{Spacing x levels equally instead of using quantiles}

\item{smooth}{A multiplier for how much smoother you want the sampling of the levels to be.}

\item{posterior}{The way the posterior is visualized: defaults to nothing, but can do "CI" for the credible interval or "trace" for traceplots}

\item{xind}{A list of the variables for which you want to run the partials. Defaults to doing all of them.}
}
\value{
Returns a nice plot
}
\description{
I don't really like the partial visualizations in dbarts so I made some of my own, with a little more customization options.
}
