% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ri.R
\name{plot.ri}
\alias{plot.ri}
\title{A plot tool for riBART models}
\usage{
plot.ri(model, temporal = TRUE)
}
\arguments{
\item{model}{Object of class rbart}

\item{temporal}{Is this yearly data (line plot) or categorical}
}
\value{
Returns a plot
}
\description{
Tiny code for tiny task.
}
