% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bababooey.R
\name{predict.bart.raster}
\alias{predict.bart.raster}
\title{Spatial prediction using BART}
\usage{
\method{predict}{bart.raster}(model, inputstack)
}
\arguments{
\item{model}{Dude that's a model. It better be a freakin lBART model dude}

\item{inputstack}{That's your rasterStack, boss}
}
\description{
I got really sick of trying to do prediction using BART for species distribution models so here's a so-far-one-function package that fixes this. Bababooey
}
