% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dbarts.R
\name{predict.dbart.raster}
\alias{predict.dbart.raster}
\title{Spatial prediction using dbarts}
\usage{
predict.dbart.raster(model, inputstack, ci = FALSE, plots = FALSE)
}
\arguments{
\item{model}{model object from dbarts package}

\item{inputstack}{A RasterStack of environmental predictors; make sure they have the same names as the predictors that went into the model object.}
}
\description{
Spatial prediction of a BART model
}
