% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict2.bart}
\alias{predict2.bart}
\title{predict() for spatial use of BART models}
\usage{
predict2.bart(object, inputstack, splitby = 1, quantiles = c())
}
\arguments{
\item{object}{A BART model object generated by the dbarts package}

\item{inputstack}{An object of class RasterStack}

\item{splitby}{If set to a value higher than 1, will split your dataset into approximately n divisible chunks}

\item{quantiles}{Include the extraction of quantiles (e.g. 5\% and 95\% credible interval) from the posterior}
}
\description{
This is gonna be messy from a development standpoint
}
