% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retune.R
\name{retune}
\alias{retune}
\title{Wrapper for xbart}
\usage{
retune(object, reps = 10)
}
\arguments{
\item{object}{A classic BART model (not rbart) from the dbarts package}

\item{reps}{How many tries with each parameter combination}
}
\description{
STILL IN DEVELOPMENT. New function to automate a little tuning of priors. Currently doesn't let you set the levels. Model goes in, model comes back out with slightly different priors.
}
