% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartials.R
\name{spartial}
\alias{spartial}
\title{Spatial partial dependence plots}
\usage{
spartial(
  model,
  envs,
  x.vars = NULL,
  equal = FALSE,
  smooth = 1,
  transform = TRUE,
  save = TRUE
)
}
\arguments{
\item{model}{A dbarts model object}

\item{envs}{The rasterStack of environmental variables, OR (this is new functionality and mostly untested, but important if you're doing tSDMs) a list of rasterstacks (e.g., climate layers by year). If it's a list, this will return a rasterstack of spartials ordered by that list, then by variable (e.g., x1-stack1, x1-stack2, x2-stack1, x2-stack2)}

\item{x.vars}{The particular x variables you want to map}

\item{equal}{Equal spacing on x breaks (versus quantiles).}

\item{smooth}{Smoothing factor for the x breaks (works like partials).}

\item{transform}{Backtransform the pnorm or not}

\item{save}{Turn this on if you want to save the outputs as a RasterStack, for nicer use/plotting elsewhere.}
}
\value{
Returns a nice plot
}
\description{
Project partial dependence surfaces in the real, geography-having world: see the spatial signal of any given variable's contribution to the model predictions.
}
