% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bart}
\alias{summary.bart}
\title{AUC for BART object}
\usage{
\method{summary}{bart}(object, plots = TRUE)
}
\arguments{
\item{model}{A BART model object generated by the dbarts package}
}
\description{
An all-purpose model diagnostic, that mostly generates an AUC value and receiver operator curve. It also prints the optimal threshold based on the true skill statistic (and return the associated Type I and II error rates).
}
