% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{var.imp}
\alias{var.imp}
\title{Variable contributions}
\usage{
var.imp(model)
}
\arguments{
\item{model}{Dude that's a model. It better be a freakin lBART model dude}
}
\description{
Crude but: something?
}
