% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC for bart.R
\name{bart.auc}
\alias{bart.auc}
\title{AUC for BART object}
\usage{
bart.auc(model, true.vector)
}
\arguments{
\item{model}{Dude that's a model. It better be a freakin lBART model dude}
}
\description{
Just for object from BART package
}
