% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartflex.R
\name{bart.flex}
\alias{bart.flex}
\title{Wrapper for flexible riBART calls}
\usage{
bart.flex(
  x.data,
  y.data,
  ri.data = NULL,
  y.name = NULL,
  ri.name = NULL,
  n.trees = 200,
  k = 2,
  power = 2,
  base = 0.95
)
}
\arguments{
\item{x.data}{A data frame of covariates}

\item{y.data}{A vector of outcomes (1/0)}

\item{ri.data}{Random intercept data}
}
\description{
A little internal piece of code that allows you to easily drop random intercepts in. It's a convenience tool I don't suggest using - it's a shortcut for if you feel very comfortable with the pipeline.
}
