% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dbarts.R
\name{bart.map}
\alias{bart.map}
\title{Spatial prediction using dbarts}
\usage{
bart.map(model, inputstack, ci = FALSE, plots = FALSE)
}
\arguments{
\item{model}{model object from dbarts package}

\item{inputstack}{A RasterStack of environmental predictors; make sure they have the same names as the predictors that went into the model object.}
}
\description{
Spatial prediction of a BART model
}
