% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{partial}
\alias{partial}
\title{Better, nicer, friendlier partials}
\usage{
partial(model, x.vars = NULL, equal = FALSE, smooth = 1, ci = TRUE,
  trace = TRUE, transform = TRUE, panels = FALSE)
}
\arguments{
\item{model}{A dbarts model object}

\item{x.vars}{A list of the variables for which you want to run the partials. Defaults to doing all of them.}

\item{equal}{Spacing x levels equally instead of using quantiles}

\item{smooth}{A multiplier for how much smoother you want the sampling of the levels to be.}

\item{posterior}{The way the posterior is visualized: defaults to nothing, but can do "CI" for the credible interval or "trace" for traceplots}
}
\value{
Returns a nice plot
}
\description{
I don't really like the partial visualizations in dbarts so I made some of my own, with a little more customization options.
}
\examples{
f <- function(x) { return(0.5 * x[,1] + 2 * x[,2] * x[,3]) - 5*x[,4] }
sigma <- 0.2
n <- 100
x <- matrix(2 * runif(n * 3) -1, ncol = 3)
x <- data.frame(x)
x[,4] <- rbinom(100, 1, 0.3)
colnames(x) <- c('rob', 'hugh', 'ed', 'phil')
Ey <- f(x)
y  <- rnorm(n, Ey, sigma)
df <- data.frame(y, x)
set.seed(99)

bartFit <- bart(y ~ rob + hugh + ed + phil, df,
               keepevery = 10, ntree = 100, keeptrees = TRUE)

partial(bartFit, x.vars='hugh', trace=TRUE, ci=TRUE)
partial(bartFit, x.vars='hugh', equal=TRUE, trace=TRUE, ci=TRUE)
partial(bartFit, x.vars='hugh', equal=TRUE, smooth=10, trace=TRUE, ci=TRUE)

partial(bartFit, x.vars='rob', equal=TRUE, smooth=10, trace=FALSE, ci=TRUE)
partial(bartFit, x.vars='ed', equal=TRUE, smooth=10, trace=TRUE, ci=FALSE)
partial(bartFit, equal=TRUE, smooth=10, trace=FALSE, ci=TRUE, panel=TRUE)

}
