% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bart.R
\name{predict.bart.raster}
\alias{predict.bart.raster}
\title{Spatial prediction using BART}
\usage{
\method{predict}{bart.raster}(model, inputstack, cores = 3)
}
\arguments{
\item{model}{lBART or gBART model from the BART package}

\item{inputstack}{That's your rasterStack. check var names}
}
\description{
I got really sick of trying to do prediction using BART for species distribution models so here's a so-far-one-function package that fixes this. Bababooey
}
