% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dbarts.R
\name{predict.dbart.raster}
\alias{predict.dbart.raster}
\title{Spatial prediction using dbarts}
\usage{
predict.dbart.raster(model, inputstack, cores = 3)
}
\arguments{
\item{model}{model object from dbarts package}

\item{inputstack}{That's your rasterStack. check var names}
}
\description{
dbarts analog code for prediction
}
