% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict2.bart}
\alias{predict2.bart}
\title{predict() for spatial use of BART models}
\usage{
predict2.bart(object, x.layers, quantiles = c(), ri.data = NULL,
  ri.name = NULL, ri.pred = FALSE, splitby = 1, quiet = FALSE)
}
\arguments{
\item{object}{A BART model objector riBART model object generated by the dbarts package}

\item{x.layers}{An object of class RasterStack}

\item{quantiles}{Include the extraction of quantiles (e.g. 5\% and 95\% credible interval) from the posterior}

\item{ri.data}{If 'object' is a riBART model, this gives either one consistent value (e.g. a prediction year) or a RasterLayer for the random effect}

\item{ri.name}{The name of the random intercept in the riBART model}

\item{ri.pred}{Should the random intercept be *included* in the prediction value or dropped? Defaults to FALSE (treats the random intercept as noise to be excluded)}

\item{splitby}{If set to a value higher than 1, will split your dataset into approximately n divisible chunks}

\item{quiet}{No progress bars}
}
\description{
A predict() wrapper for combining BART models with spatial input data, to generate a Raster or RasterStack of predicted outputs. This now includes the ability to predict from random intercept models, which can be used to deal with clustering in space and time of outcome variables!
}
