% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartials.R
\name{spartial}
\alias{spartial}
\title{Spatial partial dependence plots}
\usage{
spartial(model, envs, x.vars = NULL, equal = FALSE, smooth = 1,
  transform = TRUE, save = TRUE)
}
\arguments{
\item{model}{A dbarts model object}

\item{envs}{The rasterStack of environmental variables}

\item{x.vars}{The particular x variables you want to map}

\item{equal}{Equal spacing on x breaks (versus quantiles).}

\item{smooth}{Smoothing factor for the x breaks (works like partials).}

\item{transform}{Backtransform the pnorm or not}

\item{save}{Turn this on if you want to save the outputs as a RasterStack, for nicer use/plotting elsewhere.}
}
\value{
Returns a nice plot
}
\description{
Why don't ecologists project partial dependence surfaces in the real, geography-having world? Now they can. That's probably good right? Am I doing good? Father?
}
\examples{
library(embarcadero)
library(velox)
set.seed(12345)

data(ticks)
data(covs)
mod <- SpatialPointsDataFrame(ticks[,3:4],data.frame(ticks[,1]))
names(mod@data) <- 'Presence'
tmp=rasterize(mod, covs[[1]], field="Presence", fun="min")
pts.sp1=rasterToPoints(tmp, fun=function(x){x>0})
pres.cov <- raster::extract(covs, pts.sp1[,1:2])
absence <- randomPoints(covs,nrow(ticks)) 
abs.cov <- raster::extract(covs, absence)
pres.cov <- data.frame(pres.cov); pres.cov$tick <- 1
abs.cov <- data.frame(abs.cov); abs.cov$tick <- 0
all.cov <- rbind(pres.cov, abs.cov)
all.cov <- all.cov[complete.cases(all.cov),]

first.model <- bart(all.cov[,1:11], all.cov[,'tick'], keeptrees=TRUE)

s1 <- spartial(first.model, covs, x.vars=c('bio1'),
               equal = TRUE, smooth=5)

}
