% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartials.R
\name{spartial}
\alias{spartial}
\title{Spatial partial dependence plots}
\usage{
spartial(model, envs, x.vars = NULL, equal = TRUE, smooth = 1,
  ci = TRUE, transform = TRUE, save = FALSE)
}
\arguments{
\item{model}{A dbarts model object}

\item{x.vars}{A list of the variables for which you want to run the partials. Defaults to doing all of them.}

\item{equal}{Spacing x levels equally instead of using quantiles}

\item{smooth}{A multiplier for how much smoother you want the sampling of the levels to be.}

\item{posterior}{The way the posterior is visualized: defaults to nothing, but can do "CI" for the credible interval or "trace" for traceplots}
}
\value{
Returns a nice plot
}
\description{
Why don't ecologists project partial dependence surfaces in the real, geography-having world? Now they can. That's probably good right? Am I doing good? Father?
}
\examples{
library(embarcadero)
library(velox)
set.seed(12345)

}
