% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bart}
\alias{summary.bart}
\title{Summary of a BART object}
\usage{
\method{summary}{bart}(object, plots = TRUE)
}
\arguments{
\item{plots}{Generate diagnostic plots}

\item{model}{A BART model object generated by the dbarts package}
}
\description{
An all-purpose model summary tool, which reports the model call, an AUC value (the area under the receiver-operator curve), and the optimal threshold based on the true skill statistic (and return the associated Type I and II error rates). 

There are also four diagnostic plots generated (plots=FALSE to turn off): the receiver-operator curve; a histogram of fitted values; the threshold-performance curve, which reports the true skill statistic for a given cutoff, and has a line at the optimal value; and a dotplot showing the fitted values split by the threshold and the true y values in the data.
}
