% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovariable.R
\name{variable.step}
\alias{variable.step}
\title{Automatic variable selection}
\usage{
variable.step(x.data, y.data, n.trees = 10, iter = 50)
}
\arguments{
\item{n.trees}{How many trees to use in the variable set reduction. Should be a SMALL number for optimal performance (10 or 20 trees)}

\item{iter}{How many BART models to run for each iteration of the stepwise reduction}

\item{xdata}{A data frame of covariates}

\item{ydata}{A vector of outcomes (1/0)}
}
\value{
Returns a list of the best variable set, and does some diagnostic plots along the way
}
\description{
The automated stepwise variable set reduction algorithm.
}
