% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varplot_importance.R
\name{varimp}
\alias{varimp}
\title{Variable importance plot}
\usage{
varimp(model, names, plots = FALSE)
}
\arguments{
\item{model}{The dbarts model object}

\item{names}{The variable names (not conserved in dbarts necessarily)}

\item{plots}{Turn this on for a nice variable contribution plot}
}
\description{
Variable importance, as measured in the proportion of total branches used for a given variable.
}
